requirements.txt
django~=5.1.3
psycopg~=3.2.3
python-decouple~=3.8
python-dotenv~=1.0.1
django-ckeditor~=6.7.1